pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Dope Soft Network Vulnerability Scanner"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "python-nmap>=0.7.1",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
]
